EY-3967: Trekkjer ut manuell revurdering til eiga service
RevurderingService-klassa vart brukt frå både automatisk og manuell revurdering. Det gjorde det litt forvirrande, særleg med tankje på at automatisk behandling allereie er skilt ut i si eiga service. Innfører derfor service-klasse også for manuell, og flyttar det openbart manuelt-relaterte dit
