+++ apps/etterlatte-behandling/src/main/kotlin/Application.kt
@@ -174,7 +174,10 @@ private fun Route.settOppRoutes(applicationContext: ApplicationContext) {
         sakService = applicationContext.sakService,
     )
     vedtaksbehandlingRoutes(vedtaksbehandlingService = applicationContext.vedtaksbehandlingService)
-    revurderingRoutes(revurderingService = applicationContext.revurderingService)
+    revurderingRoutes(
+        revurderingService = applicationContext.revurderingService,
+        manuellRevurderingService = applicationContext.manuellRevurderingService,
+    )
     omregningRoutes(omregningService = applicationContext.omregningService)
     migreringRoutes(migreringService = applicationContext.migreringService)
     bosattUtlandRoutes(bosattUtlandService = applicationContext.bosattUtlandService)
