+++ apps/etterlatte-behandling/src/main/kotlin/behandling/revurdering/RevurderingService.kt
@@ -1,6 +1,5 @@
 package no.nav.etterlatte.behandling.revurdering
 
-import io.ktor.server.plugins.BadRequestException
 import kotlinx.coroutines.runBlocking
 import no.nav.etterlatte.behandling.BehandlingDao
 import no.nav.etterlatte.behandling.BehandlingHendelserKafkaProducer
@@ -8,7 +7,6 @@ import no.nav.etterlatte.behandling.BehandlingService
 import no.nav.etterlatte.behandling.GrunnlagService
 import no.nav.etterlatte.behandling.aktivitetsplikt.AktivitetspliktDao
 import no.nav.etterlatte.behandling.aktivitetsplikt.AktivitetspliktKopierService
-import no.nav.etterlatte.behandling.domain.Behandling
 import no.nav.etterlatte.behandling.domain.OpprettBehandling
 import no.nav.etterlatte.behandling.domain.Revurdering
 import no.nav.etterlatte.behandling.domain.toBehandlingOpprettet
@@ -37,7 +35,6 @@ import no.nav.etterlatte.libs.common.oppgave.OppgaveIntern
 import no.nav.etterlatte.libs.common.oppgave.OppgaveKilde
 import no.nav.etterlatte.libs.common.oppgave.OppgaveType
 import no.nav.etterlatte.libs.common.tidspunkt.Tidspunkt
-import no.nav.etterlatte.libs.common.tidspunkt.toLocalDatetimeUTC
 import no.nav.etterlatte.libs.ktor.token.Saksbehandler
 import no.nav.etterlatte.oppgave.OppgaveService
 import org.slf4j.LoggerFactory
@@ -130,131 +127,6 @@ class RevurderingService(
             throw MaksEnAktivOppgavePaaBehandling(sakId)
         }
     }
-
-    fun opprettManuellRevurderingWrapper(
-        sakId: Long,
-        aarsak: Revurderingaarsak,
-        paaGrunnAvHendelseId: String?,
-        paaGrunnAvOppgaveId: String? = null,
-        begrunnelse: String?,
-        fritekstAarsak: String? = null,
-        saksbehandler: Saksbehandler,
-    ): Revurdering {
-        if (!aarsak.kanBrukesIMiljo()) {
-            throw RevurderingaarsakIkkeStoettet(aarsak)
-        }
-        val paaGrunnAvHendelseUuid =
-            try {
-                paaGrunnAvHendelseId?.let { UUID.fromString(it) }
-            } catch (e: Exception) {
-                throw BadRequestException(
-                    "$aarsak har en ugyldig hendelse id for sakid" +
-                        " $sakId. " +
-                        "Hendelsesid: $paaGrunnAvHendelseId",
-                )
-            }
-
-        val paaGrunnAvOppgaveUuid =
-            try {
-                paaGrunnAvOppgaveId?.let { UUID.fromString(it) }
-            } catch (e: Exception) {
-                throw BadRequestException("Ugyldig oppgaveId $paaGrunnAvOppgaveId (sakid=$sakId).")
-            }
-
-        maksEnOppgaveUnderbehandlingForKildeBehandling(sakId)
-        val forrigeIverksatteBehandling =
-            behandlingService.hentSisteIverksatte(sakId)
-                ?: throw RevurderingManglerIverksattBehandling(sakId)
-
-        val sakType = forrigeIverksatteBehandling.sak.sakType
-        if (!aarsak.gyldigForSakType(sakType)) {
-            throw BadRequestException("$aarsak er ikke støttet for $sakType")
-        }
-        kanOppretteRevurderingForAarsak(aarsak)
-        return opprettManuellRevurdering(
-            sakId = forrigeIverksatteBehandling.sak.id,
-            forrigeBehandling = forrigeIverksatteBehandling,
-            revurderingAarsak = aarsak,
-            paaGrunnAvHendelse = paaGrunnAvHendelseUuid,
-            paaGrunnAvOppgave = paaGrunnAvOppgaveUuid,
-            begrunnelse = begrunnelse,
-            fritekstAarsak = fritekstAarsak,
-            saksbehandler = saksbehandler,
-            opphoerFraOgMed =
-                if (aarsak !=
-                    Revurderingaarsak.REVURDERE_ETTER_OPPHOER
-                ) {
-                    forrigeIverksatteBehandling.opphoerFraOgMed
-                } else {
-                    null
-                },
-        )
-    }
-
-    private fun kanOppretteRevurderingForAarsak(aarsak: Revurderingaarsak) {
-        if (aarsak == Revurderingaarsak.OMGJOERING_ETTER_KLAGE) {
-            throw UgyldigForespoerselException(
-                code = "OMGJOERING_KLAGE_MAA_OPPRETTES_FRA_OPPGAVE",
-                detail = "Omgjøring etter klage må opprettes fra en omgjøringsoppgave for å koble omgjøringen til klagen riktig.",
-            )
-        }
-    }
-
-    private fun opprettManuellRevurdering(
-        sakId: Long,
-        forrigeBehandling: Behandling,
-        revurderingAarsak: Revurderingaarsak,
-        paaGrunnAvHendelse: UUID?,
-        paaGrunnAvOppgave: UUID?,
-        begrunnelse: String?,
-        fritekstAarsak: String?,
-        saksbehandler: Saksbehandler,
-        opphoerFraOgMed: YearMonth? = null,
-    ): Revurdering =
-        forrigeBehandling.let {
-            val persongalleri = runBlocking { grunnlagService.hentPersongalleri(forrigeBehandling.id) }
-            val triggendeOppgave = paaGrunnAvOppgave?.let { oppgaveService.hentOppgave(it) }
-
-            opprettRevurdering(
-                sakId = sakId,
-                persongalleri = persongalleri,
-                forrigeBehandling = forrigeBehandling.id,
-                mottattDato = Tidspunkt.now().toLocalDatetimeUTC().toString(),
-                prosessType = Prosesstype.MANUELL,
-                kilde = Vedtaksloesning.GJENNY,
-                revurderingAarsak = revurderingAarsak,
-                virkningstidspunkt = null,
-                utlandstilknytning = forrigeBehandling.utlandstilknytning,
-                boddEllerArbeidetUtlandet = forrigeBehandling.boddEllerArbeidetUtlandet,
-                begrunnelse = begrunnelse ?: triggendeOppgave?.merknad,
-                fritekstAarsak = fritekstAarsak,
-                saksbehandlerIdent = saksbehandler.ident,
-                frist = triggendeOppgave?.frist,
-                paaGrunnAvOppgave = paaGrunnAvOppgave,
-                opphoerFraOgMed = opphoerFraOgMed,
-            ).oppdater()
-                .also { revurdering ->
-                    if (paaGrunnAvHendelse != null) {
-                        grunnlagsendringshendelseDao.settBehandlingIdForTattMedIRevurdering(
-                            paaGrunnAvHendelse,
-                            revurdering.id,
-                        )
-                        try {
-                            oppgaveService.ferdigStillOppgaveUnderBehandling(
-                                paaGrunnAvHendelse.toString(),
-                                OppgaveType.VURDER_KONSEKVENS,
-                                saksbehandler,
-                            )
-                        } catch (e: Exception) {
-                            logger.error(
-                                "Kunne ikke ferdigstille oppgaven til hendelsen på grunn av feil, " +
-                                    "men oppgave er ikke i bruk i miljø så feilen svelges.",
-                                e,
-                            )
-                        }
-                    }
-                }
-        }
 
     private fun behandlingErAvTypenRevurderingOgKanEndres(behandlingId: UUID) {
         val behandling =
